Move v1-exclusive code to appropriate module
Code that was only for v1 and NOT v2 was in the common folder.
